# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/imports/accounts/{accountId}/preview": {
      "post": {
        "operationId": "ImportsController_preview",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response. All records are demo data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "400": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "CSV Imports"
        ]
      }
    },
    "/api/v1/imports": {
      "get": {
        "operationId": "ImportsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 10000,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response. All records are demo data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSummaryPage"
                }
              }
            }
          },
          "400": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "CSV Imports"
        ]
      }
    },
    "/api/v1/imports/{id}": {
      "get": {
        "operationId": "ImportsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response. All records are demo data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "400": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "CSV Imports"
        ]
      }
    },
    "/api/v1/imports/{id}/commit": {
      "post": {
        "operationId": "ImportsController_commit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitImportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response. All records are demo data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "400": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Consistent error envelope; see error.code and requestId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "CSV Imports"
        ]
      }
    },
...
  },
  "info": {
    "title": "RASID API",
    "description": "Arabic-first backend portfolio. Synthetic/manual/sanitized demo data only. All money values are decimal strings. No banking, payments, investment execution or financial advice. See docs/FRONTEND_HANDOFF.md for browser integration and response examples.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "rasid_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rasid_refresh"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CommitImportDto": {
        "type": "object",
        "properties": {
          "acknowledgeRejectedRows": {
            "type": "object",
            "default": false,
            "description": "Required true when the preview contains INVALID rows; only ACCEPTED rows will be committed."
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "messageAr": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {}
              }
            },
            "required": [
              "code",
              "message",
              "messageAr"
            ]
          },
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "error",
          "requestId",
          "timestamp"
        ]
      },
...
      "ImportRow": {
        "type": "object",
        "properties": {
          "row": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACCEPTED",
              "INVALID",
              "DUPLICATE"
            ]
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "record": {
            "type": "object",
            "properties": {
              "postedAt": {
                "type": "string",
                "format": "date"
              },
              "amount": {
                "type": "string",
                "example": "42.50",
                "description": "Exact decimal string, never a JSON floating-point number."
              },
              "direction": {
                "type": "string",
                "enum": [
                  "INCOME",
                  "EXPENSE"
                ]
              },
              "merchant": {
                "type": "string"
              },
              "categoryId": {
                "type": "string",
                "format": "uuid",
                "nullable": true
              },
              "reference": {
                "type": "string",
                "nullable": true
              },
              "fingerprint": {
                "type": "string"
              }
            },
            "required": [
              "postedAt",
              "amount",
              "direction",
              "merchant",
              "categoryId",
              "reference",
              "fingerprint"
            ]
          }
        },
        "required": [
          "row",
          "status",
          "errors"
        ]
      },
      "ImportSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PREVIEW",
              "COMMITTED"
            ]
          },
          "filename": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          },
          "committedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "result": {
            "type": "object",
            "properties": {
              "inserted": {
                "type": "integer"
              },
              "duplicates": {
                "type": "integer"
              },
              "invalid": {
                "type": "integer"
              }
            },
            "required": [
              "inserted",
              "duplicates",
              "invalid"
            ],
            "nullable": true
          },
          "summary": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer"
              },
              "accepted": {
                "type": "integer"
              },
              "duplicates": {
                "type": "integer"
              },
              "invalid": {
                "type": "integer"
              }
            },
            "required": [
              "total",
              "accepted",
              "duplicates",
              "invalid"
            ]
          }
        },
        "required": [
          "id",
          "createdAt",
          "userId",
          "accountId",
          "status",
          "filename",
          "contentHash",
          "committedAt",
          "result",
          "summary"
        ]
      },
      "ImportPreview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PREVIEW",
              "COMMITTED"
            ]
          },
          "filename": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          },
          "committedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "result": {
            "type": "object",
            "properties": {
              "inserted": {
                "type": "integer"
              },
              "duplicates": {
                "type": "integer"
              },
              "invalid": {
                "type": "integer"
              }
            },
            "required": [
              "inserted",
              "duplicates",
              "invalid"
            ],
            "nullable": true
          },
          "summary": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer"
              },
              "accepted": {
                "type": "integer"
              },
              "duplicates": {
                "type": "integer"
              },
              "invalid": {
                "type": "integer"
              }
            },
            "required": [
              "total",
              "accepted",
              "duplicates",
              "invalid"
            ]
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportRow"
            }
          }
        },
        "required": [
          "id",
          "createdAt",
          "userId",
          "accountId",
          "status",
          "filename",
          "contentHash",
          "committedAt",
          "result",
          "summary",
          "rows"
        ]
      },
...
      "ImportSummaryPage": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportSummary"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}